requirements.txt
opencv-python>=4.8
numpy>=1.24
scipy>=1.10
pyyaml>=6.0
